requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
binutils==0.3.dev1
cachetools==6.2.0
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.1
Choco==1.0.5
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.2
daphne==4.2.1
distlib==0.3.8
Django==5.0.6
django-allauth==65.6.0
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-phonenumber-field==8.0.0
django-recaptcha==4.1.0
django-redis==6.0.0
django-sms==0.7.0
django-timezone-field==7.1
django-webpush==0.3.6
dotenv==0.9.9
filelock==3.15.3
frozenlist==1.5.0
google-auth==2.41.1
http_ece==1.2.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jwt==1.3.1
kombu==5.5.4
MarkupSafe==2.1.5
msgpack==1.1.1
multidict==6.2.0
oauthlib==3.2.2
packaging==25.0
phonenumbers==9.0.2
pillow==11.1.0
platformdirs==4.2.2
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
py-vapid==1.9.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyfcm==2.1.0
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyotp==2.9.0
pyserial==3.5
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pywebpush==2.1.0
redis==6.4.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
sms==0.4
sqlparse==0.5.0
twilio==9.5.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
virtualenv==20.26.2
wcwidth==0.2.13
yarl==1.18.3
zope.interface==7.2
